Rename directory to fit convention
