+++ graylog2-web-interface/src/actions/content-packs/ContentPacksActions.jsx
