+++ graylog2-web-interface/src/injection/CombinedProvider.js
@@ -12,7 +12,7 @@ class CombinedProvider {
       CodecTypes: () => require('actions/codecs/CodecTypesActions'),
       Configuration: () => require('actions/configurations/ConfigurationActions'),
       ConfigurationBundles: () => require('actions/configuration-bundles/ConfigurationBundlesActions'),
-      ContentPacks: () => require('actions/content_packs/ContentPacksActions'),
+      ContentPacks: () => require('actions/content-packs/ContentPacksActions'),
       Dashboards: () => require('actions/dashboards/DashboardsActions'),
       Decorators: () => require('actions/decorators/DecoratorsActions'),
       Deflector: () => require('actions/indices/DeflectorActions'),
